LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssSkewY.tentative-expected.txt


FAIL Constructing a CSSSkewY with a keyword throws a TypeError assert_throws_js: function "() => new CSSSkewY(value)" threw object "ReferenceError: Can't find variable: CSSSkewY" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewY with a double throws a TypeError assert_throws_js: function "() => new CSSSkewY(value)" threw object "ReferenceError: Can't find variable: CSSSkewY" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewY with a unitless zero throws a TypeError assert_throws_js: function "() => new CSSSkewY(value)" threw object "ReferenceError: Can't find variable: CSSSkewY" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewY with a string angle throws a TypeError assert_throws_js: function "() => new CSSSkewY(value)" threw object "ReferenceError: Can't find variable: CSSSkewY" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewY with a number CSSUnitValue throws a TypeError assert_throws_js: function "() => new CSSSkewY(value)" threw object "ReferenceError: Can't find variable: CSSSkewY" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewY with a time dimension CSSUnitValue throws a TypeError assert_throws_js: function "() => new CSSSkewY(value)" threw object "ReferenceError: Can't find variable: CSSSkewY" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewY with a CSSMathValue of length type throws a TypeError assert_throws_js: function "() => new CSSSkewY(value)" threw object "ReferenceError: Can't find variable: CSSSkewY" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSSkewY.ay with a keyword throws a TypeError Can't find variable: CSSSkewY
FAIL Updating CSSSkewY.ay with a double throws a TypeError Can't find variable: CSSSkewY
FAIL Updating CSSSkewY.ay with a unitless zero throws a TypeError Can't find variable: CSSSkewY
FAIL Updating CSSSkewY.ay with a string angle throws a TypeError Can't find variable: CSSSkewY
FAIL Updating CSSSkewY.ay with a number CSSUnitValue throws a TypeError Can't find variable: CSSSkewY
FAIL Updating CSSSkewY.ay with a time dimension CSSUnitValue throws a TypeError Can't find variable: CSSSkewY
FAIL Updating CSSSkewY.ay with a CSSMathValue of length type throws a TypeError Can't find variable: CSSSkewY
FAIL CSSSkewY can be constructed from an angle CSSUnitValue Can't find variable: CSSSkewY
FAIL CSSSkewY can be constructed from a CSSMathValue of angle type Can't find variable: CSSSkewY
FAIL CSSSkewY.ay can be updated to an angle CSSUnitValue Can't find variable: CSSSkewY
FAIL CSSSkewY.ay can be updated to a CSSMathValue of angle type Can't find variable: CSSSkewY
FAIL Modifying skewY.is2D is a no-op Can't find variable: CSSSkewY
